#1230 added tag cleanup command
