+++ SoftLayer/CLI/tags/cleanup.py
@@ -1,0 +1,33 @@
+"""Removes unused Tags"""
+# :license: MIT, see LICENSE for more details.
+
+import click
+
+from SoftLayer.CLI import environment
+from SoftLayer.CLI import formatting
+from SoftLayer.exceptions import SoftLayerAPIError
+from SoftLayer.managers.tags import TagManager
+from SoftLayer import utils
+
+from pprint import pprint as pp
+# pylint: disable=unnecessary-lambda
+
+
+@click.command()
+@click.option('--dry-run', '-d', is_flag=True, default=False,
+              help="Don't delete, just show what will be deleted.")
+@environment.pass_env
+def cli(env, dry_run):
+    """Removes all empty tags."""
+
+    tag_manager = TagManager(env.client)
+    empty_tags = tag_manager.get_unattached_tags()
+    
+    for tag in empty_tags:
+        if dry_run:
+            click.secho("(Dry Run) Removing {}".format(tag.get('name')), fg='yellow')
+        else:
+            result = tag_manager.delete_tag(tag.get('name'))
+            color = 'green' if result else 'red'
+            click.secho("Removing {}".format(tag.get('name')), fg=color)
+
