+++ SoftLayer/CLI/tags/taggable.py
@@ -23,22 +23,6 @@ def cli(env):
         for resource in resources:
             table.add_row([
                 resource['resource']['id'],
-                get_resource_name(resource['resource'], tag_type['keyName'])
+                tag_manager.get_resource_name(resource['resource'], tag_type['keyName'])
             ])
         env.fout(table)
-
-
-def get_resource_name(resource, tag_type):
-    """Returns a string that names a resource"""
-    if tag_type == 'NETWORK_VLAN_FIREWALL':
-        return resource.get('primaryIpAddress')
-    elif tag_type == 'NETWORK_VLAN':
-        return "{} ({})".format(resource.get('vlanNumber'), resource.get('name'))
-    elif tag_type == 'IMAGE_TEMPLATE' or tag_type == 'APPLICATION_DELIVERY_CONTROLLER':
-        return resource.get('name')
-    elif tag_type == 'TICKET':
-        return resource.get('subjet')
-    elif tag_type == 'NETWORK_SUBNET':
-        return resource.get('networkIdentifier')
-    else:
-        return resource.get('fullyQualifiedDomainName')
